move run command under /actions
